requirements.txt
numpy
opencv-python
Pillow
matplotlib
PySide6
scipy